requirements.txt
auth0-python==3.19.0
Authlib==0.15.5
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
constants==0.6.0
cryptography==36.0.1
ecdsa==0.17.0
Flask==2.0.2
flask-auth0==0.1.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.5.1
GitHub-Flask==3.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jose==1.0.0
MarkupSafe==2.0.1
oauthlib==2.1.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python-jose==3.3.0
requests==2.27.1
requests-oauthlib==1.1.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.2
